+++ tidy3d/components/spice/types.py
@@ -1,0 +1,6 @@
+from typing import Union
+from tidy3d.components.spice.analysis.dc import TransferFunctionDC
+
+ElectricalAnalysisTypes = Union[
+    TransferFunctionDC
+]
